Do not clear the log entry queue unless the server received them
This change makes things a bit more reliable, since it will keep log
entries alive in memory until the server acknowledges that it has
received the logs.

Note that if the server is stopped while in this state, it will not be
able to save the log entries and they will be lost.